Add files via upload
This is a modified version of the original Noldnold.py file, where I removed all the self. funcions, because for this project they aren't necessary.
I will rewrite the code based on this Python file when no further updates are taken.
I should probably overwrite the original Noldnold file with this one.